+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 - Added view to basket for *Application Status History* (126577047)
 - Added *Product Limits* for an *Installation* (127177411)
 - Added validation for product limits (127177411)
+- Add flexible finance calculator to basket (121944781)
+- Do not show inputs for deposit limits if the deposit is fixed
 
 # 1.8.0
 2016-08-12